% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human_Membrane.R
\docType{data}
\name{Human_Membrane}
\alias{Human_Membrane}
\title{A list of ligand-receptor pairs of Human Membrane}
\format{a table}
\usage{
data(Human_Membrane)
}
\description{
this is a list of ligand and receptor pairs for Human Membrane
}
