% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human_NonMembrane.R
\docType{data}
\name{Human_NonMembrane}
\alias{Human_NonMembrane}
\title{A list of ligand-receptor pairs of Human_NonMembrane}
\format{a table}
\usage{
data(Human_NonMembrane)
}
\description{
this is a list of ligand and receptor pairs for Human_NonMembrane
}
