% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human_NonPlasma_Membrane.R
\docType{data}
\name{Human_NonPlasma_Membrane}
\alias{Human_NonPlasma_Membrane}
\title{A list of ligand-receptor pairs of Human_NonPlasma_Membrane}
\format{a table}
\usage{
data(Human_NonPlasma_Membrane)
}
\description{
this is a list of ligand and receptor pairs for Human_NonPlasma_Membrane
}
