% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Human_Plasma_Membrane.R
\docType{data}
\name{Human_Plasma_Membrane}
\alias{Human_Plasma_Membrane}
\title{A list of ligand-receptor pairs of Human_Plasma_Membrane}
\format{a table}
\usage{
data(Human_Plasma_Membrane)
}
\description{
this is a list of ligand and receptor pairs for Human_Plasma_Membrane
}
