% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC3_Alpha.R
\name{LRC3_CellTypeConnection}
\alias{LRC3_CellTypeConnection}
\title{Visualize the connections for a specific cell type.}
\usage{
LRC3_CellTypeConnection(LRC3_list, CellType_Index = 1,
  By_Proportion = FALSE, Combine_InOut = FALSE, TargetCellType_IN = FALSE,
  TargetCellType_OUT = FALSE, ViewTop_LRpairs = FALSE, TOP_Number = 50,
  SaveImage = FALSE)
}
\arguments{
\item{LRC3_list}{the input data generated by function LRC3_INF.}

\item{CellType_Index}{the Index of the queried cell type that users are interested. The maximum number is the number of cell types.}

\item{By_Proportion}{if \code{By_Proportion = FALSE}, view the connections based on the mean expression of cell types.
if \code{By_Proportion = TRUE}, view the connections based on the proportion of cells within a cell type that significantly express the specific ligand or receptor.}

\item{Combine_InOut}{If \code{Combine_InOut = TRUE}, it gives a figure that combines the IN and OUT signal tegether.}

\item{TargetCellType_IN}{if \code{TargetCellType_IN = TRUE},it shows the IN signal for the cell type in query.}

\item{TargetCellType_OUT}{if \code{TargetCellType_OUT = TRUE}, it shows the OUT signal for the cell type in query.}

\item{ViewTop_LRpairs}{if \code{ViewTop_LRpairs = FALSE}, show all the ligand-receptor pairs in the output figure.
if \code{ViewTop_UniqLRpairs = TRUE}, show a number of top highly expressed ligand-receptor pairs in the output figure.
The ligand-receptor pairs are ranked by their mean expression value in an decreasing order.}

\item{TOP_Number}{the number of ligand-receptor pairs to show if \code{ViewTop_LRpairs = TRUE}.}

\item{SaveImage}{if \code{SaveImage = FALSE},plot without saving the figure. if \code{SaveImage = TRUE},save the figure as pdf.}
}
\description{
This function extracts and plots connections of a specific cell type with other cell types.
For a certain cell type, ligand to receptor is defined as OUT signal and receptor to ligand is defined as IN signal.
By setting parameters, users can chose the type of signals to view.
The ligand-receptor pairs are ranked by their mean expression value.
}
