% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC3_Alpha.R
\name{LRC3_Connection}
\alias{LRC3_Connection}
\title{Visualize the connections between different cell types.}
\usage{
LRC3_Connection(LRC3_list, IndexOfPlot = 1, By_Proportion = FALSE,
  MultiFigure = FALSE, PlotResults = TRUE, ReturnList = FALSE,
  SaveImage = FALSE)
}
\arguments{
\item{LRC3_list}{the input data generated by function LRC3_INF.}

\item{IndexOfPlot}{the Index of the plot.
if \code{IndexOfPlot = 1},view connections with ligand-receptor pairs used more than once among cell types.
if \code{IndexOfPlot = 2},view the usage frequencies of each ligand or receptor by each cell type.
if \code{IndexOfPlot = 3},view the connections with ligand-receptor pairs used only once among cell types.}

\item{By_Proportion}{if \code{By_Proportion = FALSE}, view the connections based on the mean expression of cell types.
if \code{By_Proportion = TRUE}, view the connections based on the proportion of cells within a cell type that significantly express the specific ligand or receptor.}

\item{MultiFigure}{parameter for plotting multiple figures or ONE figure.
if \code{MultiFigure = FALSE}, show all the connections on ONE figure.
The figure would be very intense when there are a large number of connections.
if \code{MultiFigure = FALSE}, show connections on multiple figures. It gives a maximun number of 5 figures based on the total number of connections.
Apart from the last one figure, each figure shows a fixed number of 50 ligand-receptor patheays. The last figure shows all the remaining pairs.
The ligand-receptor pairs is first ranked by their usage frequency by a decreasing order.
Then pairs with same usage frequency are further ranked by their expression level by a decreasing order.}

\item{PlotResults}{if \code{PlotResults = TRUE}, plot out the figures.}

\item{ReturnList}{if \code{ReturnList = TRUE},return the results as a list.}

\item{SaveImage}{if \code{SaveImage = FALSE},plot without save the figure. if \code{SaveImage = TRUE},save the figure as pdf.}
}
\value{
LRC3_Connection_list a list that contains three elements.
1st is the table for SharedLRpairs
2nd is the table for UniqueLRpairs
3rd is the Total LRpairs Ranked by mean, including both shared and unique LRpairs
}
\description{
This function extracts and plots connections between cell types using different ligand-receptor pathways
based on the usage frequencies and expression level of the ligand-receptor pairs.
Ligand-receptor pairs that have been used for more than 1 times or only 1 time among cell types are shown seperately.
IndexNumber 1 shows connections of ligand-receptor pairs used more than 1 times. By choosing different parameters,
users can choose to view connections based on the mean expression level of a cell type,
or proportion of cells in a cell type that significantly express the specific pathways.
IndexNumber 2 shows the usage frequencies of each specific ligand or receptor by each cell type.
Due to significant test, some connections of certain pathways between cell types are removed during calculation.
IndexNumber 3 shows the connection of ligand-receptor pairs used only once among cell types.
By setting the parameter, users can view connections based on mean expression of a cell type,
or proportion of cells in a cell type that significantly express the specific ligands and receptors.
}
