% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC3_Alpha.R
\name{LRC3_ContactMatrixPlot}
\alias{LRC3_ContactMatrixPlot}
\title{The Contact Matrix Plot}
\usage{
LRC3_ContactMatrixPlot(LRC3_list, IndexNumber, Cluster = FALSE,
  SaveImage = FALSE, FileName = "Contactmatrix")
}
\arguments{
\item{LRC3_list}{the input data generated by function LRC3_INF.}

\item{IndexNumber}{the Index of the plot.}

\item{Cluster}{if \code{Cluster =FALSE}, the visualization of the result will not be clustered.}

\item{SaveImage}{if \code{SaveImage =FALSE}, save image as pdf file.}

\item{FileName}{if \code{SaveImage =TRUE}, give the name to the pdf file.}
}
\description{
This function performs the Plot on the contacts between different cell types.
It gives 4 different plots in total.
IndexNumber 1 is the matrix showing the contact number between different cell types in two direction: ligand to receptor(OUT), receptor to ligand(IN).
IndexNumber 2 is the scaled matrix showing the scaled contact number between different cell types. The scaled number is the ratio of the expression level of that gene to the threshold.
IndexNumber 3 is the undirectional contact matrix,which add the In and OUT signals between two cell types.
IndexNumber 4 is the undirectional scaled contact matrix.
Both one column and one row in the plot represent one cell type,respectively.
The color in the plot represents the contact number.
}
