% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC3_Alpha.R
\name{LRC3_INF}
\alias{LRC3_INF}
\title{Get a list of results.}
\usage{
LRC3_INF(data_list, PvalAsThreshold = TRUE, FixedThreshold = 1,
  MultiThreshold = FALSE, PercentForPval = 0.95, PercentForPtest = 0.95,
  DefaultLRlist = TRUE, Species = c("Human", "Mouse"),
  LR_pairs = LR_pairs, ncore = 2)
}
\arguments{
\item{data_list}{the input data generated by function PrepareData.}

\item{PvalAsThreshold}{parameter used to set the Pval as the threshold.
if \code{PvalAsThreshold = TRUE},the algrithm will perform a random permutation process on the expression matrix,
and get a threshold from the permutation for each of the gene.
if \code{PvalAsThreshold = FALSE},the algrithm will use one fixed threshold for all the genes.The defalt is TRUE.}

\item{FixedThreshold}{threshold when using a fixed value instead of Pval.
Uses should adust this threshold as needed.The defalt is set as FixedThreshold = 1.}

\item{MultiThreshold}{Boolean value to set weather use multiple threshold or only one threshold}

\item{PercentForPval}{threshold of the Pval. After the permutation on the expression matrix,
the Pval for each of the gene is chosen by this threshold from the permutation procedure.
 The default is 0.95, which means the the value of the top 5 percent.}

\item{PercentForPtest}{threshold for the P value test, the default is 0.95.}

\item{DefaultLRlist}{if \code{DefaultLRlist = TRUE}, the a defult list containning 2570 LR pairs will be used.
if \code{DefaultLRlist = FALSE}, users need to define their own parameter for of 'LR_pairs'.}

\item{Species}{select the species that users are analyzing. LRC3 could analyzing both human and mouse datasets.}

\item{LR_pairs}{the imput LRpairs users have to provide when setting the \code{DefaultLRlist = FALSE}
The user can also choose the Membrane or NonMembrane LR pairs for both Mouse and Human that curated and included in this package,
by setting \code{LR_pairs = Human_Membrane} or \code{LR_pairs = Human_NonMembrane} for Human,
and  \code{LR_pairs = Mouse_Membrane}, \code{LR_pairs = Mouse_NonMembrane} for Mouse.}

\item{ncore}{the number of cores to be used}
}
\value{
LRC3_list a list that contains 13 elements.
1st element is the contact number between different cell types.
2nd element is the scaled contact number between different cell types.
3rd element is an array containing the information of ligand-receptor pairs used between different cell types.
4th element is an array containing the scaled information of ligand-receptor pairs used between different cell types.
5th element is the ligand-receptor names that are used in the input dataset.
6th element is the expression level of ligand that are used in each cell type.
7th element is the expression level of receptor that are used in each cell type.
8th element is the scaled expression level of ligand that are used in each cell type.
9th element is the scaled the expression level of receptor that are used in each cell type.
10th element is the threshold for ligand.
11th element is the threshold for receptor.
12th element is the expression matrix ordered by cell type.
13th element is the index of cell types.
14th element is the contact number between each cell type after testing the significance of the total contact number
}
\description{
This function Performs calculations on the communication between different cell types and generate a list of results.The
result list contains 13 matrices or vectors with each gives one type of information on the input data.
The name of the 13 elements are contact_number_Remaining, scaled_contact_number_remaining,
Remaining_contactINF, Remaining_scaled_contactINF, rem_LRpairs,
LigandExpressionLevelINF, ReceptorExpressionLevelINF,
ScaledLigandExpressionLevelINF, ScaledReceptorExpressionLevelINF,
p, k, GroupedUniqueLRMatrix, GroupedNum_Colnames.
}
