% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC3_Alpha.R
\name{LRC3_LRPairsCellTypeSearch}
\alias{LRC3_LRPairsCellTypeSearch}
\title{Search the ligand-receptor pathways between two cell types.}
\usage{
LRC3_LRPairsCellTypeSearch(LRC3_list, CellType1, CellType2)
}
\arguments{
\item{LRC3_list}{the input data generated by function LRC3_INF.}

\item{CellType1}{the Index of one cell type in query. CellType1 is the cell type that significantly express ligand.}

\item{CellType2}{the Index of one cell type in query. CellType2 is the cell type that significantly express receptor.}
}
\description{
This function searches the ligand and receptor pairs that are used between two cell types.
It will give all pairs that being used between two cell types in one direction: CellType1 express ligand, CellType2 express receptor.
The ligand-receptor pairs are ranked by the signal of the pathway,
which is the mean of the two scaled expression value of both ligand and receptor in the corresponding pairwise cell types.
The scaled expression value is gained through the original mean expression value of a gene in a cell type devied by the threshold of the gene.
}
