% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC3_Alpha.R
\name{LRC3_TotalListOfLRpairs}
\alias{LRC3_TotalListOfLRpairs}
\title{Total list of ligand-receptor pairs.}
\usage{
LRC3_TotalListOfLRpairs(LRC3_list)
}
\arguments{
\item{LRC3_list}{the input data generated by function LRC3_INF.}
}
\value{
ListOfLRpairs a table contains information of ligand-receptor pairs used between pairwised cell types.
In the table, the first two colomn are the names of ligand and receptor pairs.
Signal is the mean of the two expression value of ligand and receptor in the pairwised cell types.
L_Signal is the signal for the ligand, which is the factor of mean expression value of the ligand in a cell type devided by the threshold of the ligand.
R_Signal is the signal for the receptor, which is the factor of the mean expression value of the receptor in a cell type devided by the threshold of the receptor.
L_Origin is the mean expression value of the ligand in a cell type.
R_Origin is the mean expression value of the receptor in a cell type.
L_Thresh is the threshold of the ligand.
R_Thresh is the threshold of the receptor.
Celltypes are the two interacting cell types that are using the pathway.
}
\description{
This function gives a table that contains the information of ligand-receptor pairs used between pairwised cell types.
It includes the name of the ligands and receptors, the signal of the pathways,
the original mean expression value of a gene, the threshold for a gene, and cell type names that uses the pathways.
The pairs are ranked by their signal of the pathways, which is the mean of the two expression value of ligand and receptor in the pairwised cell types.
}
