% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRpairs.R, R/Mouse_LRpairs.R
\docType{data}
\name{LRpairs}
\alias{LRpairs}
\alias{LRpairs}
\alias{Mouse_LRpairs}
\title{A list for ligand-receptor pairs}
\format{a table}
\usage{
data(LRpairs)

data(LRpairs)
}
\description{
this is a list of ligand and receptor pairs

this is a list of ligand and receptor pairs
}
\keyword{datasets}
