% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse_Membrane.R
\docType{data}
\name{Mouse_Membrane}
\alias{Mouse_Membrane}
\title{A list of ligand-receptor pairs of Mouse_Membrane}
\format{a table}
\usage{
data(Mouse_Membrane)
}
\description{
this is a list of ligand and receptor pairs for Mouse_Membrane
}
