% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse_NonMembrane.R
\docType{data}
\name{Mouse_NonMembrane}
\alias{Mouse_NonMembrane}
\title{A list of ligand-receptor pairs of Mouse_NonMembrane}
\format{a table}
\usage{
data(Mouse_NonMembrane)
}
\description{
this is a list of ligand and receptor pairs for Mouse_NonMembrane
}
