% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse_NonPlasma_Membrane.R
\docType{data}
\name{Mouse_NonPlasma_Membrane}
\alias{Mouse_NonPlasma_Membrane}
\title{A list of ligand-receptor pairs of Mouse_NonPlasma_Membrane}
\format{a table}
\usage{
data(Mouse_NonPlasma_Membrane)
}
\description{
this is a list of ligand and receptor pairs for Mouse_NonPlasma_Membrane
}
