% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse_Plasma_Membrane.R
\docType{data}
\name{Mouse_Plasma_Membrane}
\alias{Mouse_Plasma_Membrane}
\title{A list of ligand-receptor pairs of Mouse_Plasma_Membrane}
\format{a table}
\usage{
data(Mouse_Plasma_Membrane)
}
\description{
this is a list of ligand and receptor pairs for Mouse_Plasma_Membrane
}
