% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRC3_Alpha.R
\name{PrepareData}
\alias{PrepareData}
\title{Prepare the data}
\usage{
PrepareData(dd = dd, SingleCellExperimentObject = FALSE,
  log_counts = FALSE, SparseMatrix = FALSE, Species = c("Human", "Mouse"))
}
\arguments{
\item{dd}{The input expresstion matrix. The format of the data should be a dataframe or a matrix,
with rownames as the gene names, and colomn names as the cell type names.
It could also be an SingleCellExperiment objects with thecell types names stored in the “cell_type1” slot of the colData.}

\item{SingleCellExperimentObject}{if \code{SingleCellExperimentObject = TRUE},the imput data format is SingleCellExperiment object,
if \code{SingleCellExperimentObject = FALSE},the imput data format is an expression matrix.}

\item{log_counts}{parameter for 'SingleCellExperimentObject', if \code{log_counts = FALSE}, use none log normalized data,
if \code{log_counts = TRUE},use log normalized data.}

\item{SparseMatrix}{if \code{SparseMatrix = TRUE}, the SingleCellExperimentObject input data is a sparse matrix.}

\item{Species}{choosing a species of the dataset}
}
\value{
data_list an object that contain the expression matrix dd, the cell type name index, and cell type names.
}
\description{
This function is to prepare the data.
}
